add planets to scene, upgrade all planet image maps besides neptune and uranus to 8k
