+++ src/planets/bodies.js
@@ -1,0 +1,56 @@
+export const sphericalBodies = {
+  sun: {
+    radius: 695_700,
+    positionX: 0,
+    positionY: 0,
+    positionZ: 0
+  },
+  mercury: {
+    radius: 2_440,
+    positionX: 57_910_000,
+    positionY: 0,
+    positionZ: 0
+  },
+  venus: {
+    radius: 6_052,
+    positionX: 108_200_000,
+    positionY: 0,
+    positionZ: 0
+  },
+  earth: {
+    radius: 6_371,
+    positionX: 149_600_000,
+    positionY: 0,
+    positionZ: 0
+  },
+  mars: {
+    radius: 3_390,
+    positionX: 227_900_000,
+    positionY: 0,
+    positionZ: 0
+  },
+  jupiter: {
+    radius: 69_911,
+    positionX: 778_500_000,
+    positionY: 0,
+    positionZ: 0
+  },
+  saturn: {
+    radius: 58_232,
+    positionX: 1_429_400_000,
+    positionY: 0,
+    positionZ: 0
+  },
+  uranus: {
+    radius: 25_362,
+    positionX: 2_870_990_000,
+    positionY: 0,
+    positionZ: 0
+  },
+  neptune: {
+    radius: 24_622,
+    positionX: 4_498_250_000,
+    positionY: 0,
+    positionZ: 0
+  },
+};
