--- src/planets/planets.js
@@ -1,12 +1,0 @@
-import * as THREE from 'three';
-
-import setupSphereObject from '../util/setupSphereObject.js';
-import { sphericalBodies } from '../util/bodies.js';;
-
-//planet textures from https://www.solarsystemscope.com/textures/
-
-const planets = sphericalBodies
-
-console.log(sphericalBodies)
-
-export default planets;
