+++ src/util/setupSphereObject.js
@@ -3,7 +3,7 @@ import * as THREE from 'three';
 const SPHERE_WIDTH_DIVISIONS = 32;
 const SPHERE_HEIGHT_DIVISIONS = 32;
 
-const setupSphereObject = (sphereRadius, textureName, side) => {
+const setupSphereObject = (sphereRadius, textureName, side = undefined) => {
   const sphereGeometry = new THREE.SphereGeometry(sphereRadius, SPHERE_WIDTH_DIVISIONS, SPHERE_HEIGHT_DIVISIONS);
   const texture = new THREE.TextureLoader().load(`assets/${textureName}.jpg`);
